Moved navbar to general UI directory
