+++ Food_Tracker_App/app/src/main/java/com/example/foodtracker/ui/NavBar.java
@@ -1,4 +1,4 @@
-package com.example.foodtracker.ui.ingredients;
+package com.example.foodtracker.ui;
 
 import android.content.Intent;
 import android.graphics.PorterDuff;
@@ -12,6 +12,7 @@ import androidx.fragment.app.Fragment;
 
 import com.example.foodtracker.R;
 import com.example.foodtracker.model.MenuItem;
+import com.example.foodtracker.ui.ingredients.IngredientsMainScreen;
 import com.example.foodtracker.ui.mealPlan.MealPlanMainScreen;
 import com.example.foodtracker.ui.recipes.RecipesMainScreen;
 import com.example.foodtracker.ui.shoppingCart.ShoppingCartMainScreen;
