+++ Food_Tracker_App/app/src/main/java/com/example/foodtracker/ui/mealPlan/MealPlanMainScreen.java
@@ -5,7 +5,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import com.example.foodtracker.R;
 import com.example.foodtracker.model.MenuItem;
-import com.example.foodtracker.ui.ingredients.NavBar;
+import com.example.foodtracker.ui.NavBar;
 
 
 public class MealPlanMainScreen extends AppCompatActivity {
